storage: split up store.go
store.go is getting quite large and can benefit from some logical decomposition.
This PR is just code movement. It creates 6 new files in the storage package
from store.go:

    store_create_replica.go - deals with creating replicas, especially
        getOrCreateReplica().
    store_remove_replica.go - deals with removing replicas.
    store_raft.go - deals with the events from the scheduler and the raft
        transport.
    store_split.go - deals with the splitting ranges.
    store_merge.go - deals with merging ranges.
    store_send.go - Store.Send() and its helpers.

I'm also not particularly committed to splitting it up exactly
like this, just wanted to do something.

Release justification: Just code movement, better to do it before the release
for the sake of future backports.

Release note: None
