+++ pkg/storage/store.go
@@ -44,7 +44,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/storage/idalloc"
 	"github.com/cockroachdb/cockroach/pkg/storage/intentresolver"
 	"github.com/cockroachdb/cockroach/pkg/storage/raftentry"
-	"github.com/cockroachdb/cockroach/pkg/storage/stateloader"
 	"github.com/cockroachdb/cockroach/pkg/storage/tscache"
 	"github.com/cockroachdb/cockroach/pkg/storage/txnrecovery"
 	"github.com/cockroachdb/cockroach/pkg/storage/txnwait"
@@ -63,12 +62,10 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
 	"github.com/cockroachdb/cockroach/pkg/util/uuid"
-	crdberrors "github.com/cockroachdb/errors"
 	"github.com/cockroachdb/logtags"
 	"github.com/google/btree"
 	"github.com/pkg/errors"
 	"go.etcd.io/etcd/raft"
-	"go.etcd.io/etcd/raft/raftpb"
 	"golang.org/x/time/rate"
 )
 
@@ -363,19 +360,6 @@ func (rs *storeReplicaVisitor) EstimatedCount() int {
 		return rs.store.ReplicaCount()
 	}
 	return len(rs.repls) - rs.visited
-}
-
-type raftRequestInfo struct {
-	req        *RaftMessageRequest
-	respStream RaftMessageResponseStream
-}
-
-type raftRequestQueue struct {
-	syncutil.Mutex
-	infos []raftRequestInfo
-	// TODO(nvanbenschoten): consider recycling []raftRequestInfo slices. This
-	// could be done without any new mutex locking by storing two slices here
-	// and swapping them under lock in processRequestQueue.
 }
 
 // A Store maintains a map of ranges by start key. A Store corresponds
@@ -2085,758 +2069,6 @@ func (s *Store) AllocateRangeID(ctx context.Context) (roachpb.RangeID, error) {
 	return roachpb.RangeID(id), nil
 }
 
-// splitPreApply is called when the raft command is applied. Any
-// changes to the given ReadWriter will be written atomically with the
-// split commit.
-func splitPreApply(
-	ctx context.Context, eng engine.ReadWriter, split roachpb.SplitTrigger, r *Replica,
-) {
-	// Check on the RHS, we need to ensure that it exists and has a minReplicaID
-	// less than or equal to the replica we're about to initialize.
-	//
-	// The right hand side of the split was already created (and its raftMu
-	// acquired) in Replica.acquireSplitLock. It must be present here.
-	rightRepl, err := r.store.GetReplica(split.RightDesc.RangeID)
-	if err != nil {
-		log.Fatalf(ctx, "unable to find RHS replica: %+v", err)
-	}
-
-	// If the RHS is not in the split, sanity check that the LHS is currently
-	// catching up from a preemptive snapshot. A preemptive snapshot is
-	// the only reason we should here; replicas delete themselves when they
-	// apply a command which removes them from the range. The exception to that
-	// is if the DisableEagerReplicaRemoval testing flag is enabled.
-	//
-	// TODO(ajwerner): rethink DisableEagerReplicaRemoval and remove this in
-	// 20.1 after there are no more preemptive snapshots.
-	_, hasRightDesc := split.RightDesc.GetReplicaDescriptor(r.StoreID())
-	if !hasRightDesc {
-		_, lhsExists := r.Desc().GetReplicaDescriptor(r.StoreID())
-		if lhsExists {
-			log.Fatalf(ctx, "cannot process split on s%s which exists in LHS and not in RHS: %+v",
-				r.StoreID(), split)
-		}
-	}
-
-	// Check to see if we know that the RHS has already been removed from this
-	// store at the replica ID implied by the split.
-	if rightRepl.isNewerThanSplit(&split) {
-		// We're in the rare case where we know that the RHS has been removed
-		// and re-added with a higher replica ID. We know we've never processed a
-		// snapshot for the right range because up to this point it would overlap
-		// with the left and ranges cannot move rightwards.
-		//
-		// It is important to preserve the HardState because we might however have
-		// already voted at a higher term. In general this shouldn't happen because
-		// we add learners and then promote them only after we snapshot but we're
-		// going to be extra careful in case future versions of cockroach somehow
-		// promote replicas without ensuring that a snapshot has been received.
-		//
-		// Clear the user data the RHS would have inherited from the LHS due to the
-		// split and additionally clear all of the range ID local state that the
-		// split trigger writes into the RHS.
-		//
-		// Rather than specifically deleting around the data we want to preserve
-		// we read the HardState to preserve it, clear everything and write back
-		// the HardState and tombstone.
-		hs, err := rightRepl.raftMu.stateLoader.LoadHardState(ctx, eng)
-		if err != nil {
-			log.Fatalf(ctx, "failed to load hard state for removed rhs: %v", err)
-		}
-		const rangeIDLocalOnly = false
-		const mustUseClearRange = false
-		if err := clearRangeData(&split.RightDesc, eng, eng, rangeIDLocalOnly, mustUseClearRange); err != nil {
-			log.Fatalf(ctx, "failed to clear range data for removed rhs: %v", err)
-		}
-		if err := rightRepl.raftMu.stateLoader.SetHardState(ctx, eng, hs); err != nil {
-			log.Fatalf(ctx, "failed to set hard state with 0 commit index for removed rhs: %v", err)
-		}
-		if err := r.setTombstoneKey(ctx, eng, r.minReplicaID()); err != nil {
-			log.Fatalf(ctx, "failed to set tombstone for removed rhs: %v", err)
-		}
-		return
-	}
-
-	// Update the raft HardState with the new Commit value now that the
-	// replica is initialized (combining it with existing or default
-	// Term and Vote). This is the common case.
-	rsl := stateloader.Make(split.RightDesc.RangeID)
-	if err := rsl.SynthesizeRaftState(ctx, eng); err != nil {
-		log.Fatal(ctx, err)
-	}
-}
-
-// splitPostApply is the part of the split trigger which coordinates the actual
-// split with the Store. Requires that Replica.raftMu is held.
-//
-// TODO(tschottdorf): Want to merge this with SplitRange, but some legacy
-// testing code calls SplitRange directly.
-func splitPostApply(
-	ctx context.Context, deltaMS enginepb.MVCCStats, split *roachpb.SplitTrigger, r *Replica,
-) {
-	// The right hand side of the split was already created (and its raftMu
-	// acquired) in Replica.acquireSplitLock. It must be present here.
-	rightRngOrNil, err := r.store.GetReplica(split.RightDesc.RangeID)
-	if err != nil {
-		log.Fatalf(ctx, "unable to find RHS replica: %+v", err)
-	}
-	// Already holding raftMu, see above.
-	rightRngOrNil.mu.Lock()
-
-	// If we know that the RHS has already been removed at this replica ID
-	// then we also know that its data has already been removed by the preApply
-	// so we skip initializing it as the RHS of the split.
-	if rightRngOrNil.isNewerThanSplitRLocked(split) {
-		rightRngOrNil.mu.Unlock()
-		rightRngOrNil = nil
-	} else {
-		rightRng := rightRngOrNil
-		// Finish initialization of the RHS.
-		err := rightRng.initRaftMuLockedReplicaMuLocked(&split.RightDesc, r.store.Clock(), 0)
-		rightRng.mu.Unlock()
-		if err != nil {
-			log.Fatal(ctx, err)
-		}
-
-		// This initialMaxClosedValue is created here to ensure that follower reads
-		// do not regress following the split. After the split occurs there will be no
-		// information in the closedts subsystem about the newly minted RHS range from
-		// its leaseholder's store. Furthermore, the RHS will have a lease start time
-		// equal to that of the LHS which might be quite old. This means that
-		// timestamps which follow the least StartTime for the LHS part are below the
-		// current closed timestamp for the LHS would no longer be readable on the RHS
-		// after the split. It is critical that this call to maxClosed happen during
-		// the splitPostApply so that it refers to a LAI that is equal to the index at
-		// which this lease was applied. If it were to refer to a LAI after the split
-		// then the value of initialMaxClosed might be unsafe.
-		initialMaxClosed := r.maxClosed(ctx)
-		r.mu.Lock()
-		rightRng.mu.Lock()
-		// Copy the minLeaseProposedTS from the LHS.
-		rightRng.mu.minLeaseProposedTS = r.mu.minLeaseProposedTS
-		rightRng.mu.initialMaxClosed = initialMaxClosed
-		rightLease := *rightRng.mu.state.Lease
-		rightRng.mu.Unlock()
-		r.mu.Unlock()
-
-		// We need to explicitly wake up the Raft group on the right-hand range or
-		// else the range could be underreplicated for an indefinite period of time.
-		//
-		// Specifically, suppose one of the replicas of the left-hand range never
-		// applies this split trigger, e.g., because it catches up via a snapshot that
-		// advances it past this split. That store won't create the right-hand replica
-		// until it receives a Raft message addressed to the right-hand range. But
-		// since new replicas start out quiesced, unless we explicitly awaken the
-		// Raft group, there might not be any Raft traffic for quite a while.
-		err = rightRng.withRaftGroup(true, func(r *raft.RawNode) (unquiesceAndWakeLeader bool, _ error) {
-			return true, nil
-		})
-		if err != nil {
-			log.Fatalf(ctx, "unable to create raft group for right-hand range in split: %+v", err)
-		}
-
-		// Invoke the leasePostApply method to ensure we properly initialize
-		// the replica according to whether it holds the lease. This enables
-		// the txnWaitQueue.
-		rightRng.leasePostApply(ctx, rightLease, false /* permitJump */)
-	}
-	// Add the RHS replica to the store. This step atomically updates
-	// the EndKey of the LHS replica and also adds the RHS replica
-	// to the store's replica map.
-	if err := r.store.SplitRange(ctx, r, rightRngOrNil, split); err != nil {
-		// Our in-memory state has diverged from the on-disk state.
-		log.Fatalf(ctx, "%s: failed to update Store after split: %+v", r, err)
-	}
-
-	// Update store stats with difference in stats before and after split.
-	r.store.metrics.addMVCCStats(deltaMS)
-
-	now := r.store.Clock().Now()
-
-	// While performing the split, zone config changes or a newly created table
-	// might require the range to be split again. Enqueue both the left and right
-	// ranges to speed up such splits. See #10160.
-	r.store.splitQueue.MaybeAddAsync(ctx, r, now)
-	// If the range was not properly replicated before the split, the replicate
-	// queue may not have picked it up (due to the need for a split). Enqueue
-	// both the left and right ranges to speed up a potentially necessary
-	// replication. See #7022 and #7800.
-	r.store.replicateQueue.MaybeAddAsync(ctx, r, now)
-
-	if rightRngOrNil != nil {
-		r.store.splitQueue.MaybeAddAsync(ctx, rightRngOrNil, now)
-		r.store.replicateQueue.MaybeAddAsync(ctx, rightRngOrNil, now)
-		if len(split.RightDesc.Replicas().All()) == 1 {
-			// TODO(peter): In single-node clusters, we enqueue the right-hand side of
-			// the split (the new range) for Raft processing so that the corresponding
-			// Raft group is created. This shouldn't be necessary for correctness, but
-			// some tests rely on this (e.g. server.TestNodeStatusWritten).
-			r.store.enqueueRaftUpdateCheck(rightRngOrNil.RangeID)
-		}
-	}
-
-}
-
-// SplitRange shortens the original range to accommodate the new range. The new
-// range is added to the ranges map and the replicasByKey btree. origRng.raftMu
-// and newRng.raftMu must be held.
-//
-// This is only called from the split trigger in the context of the execution
-// of a Raft command. Note that rightRepl will be nil if the replica described
-// by rightDesc is known to have been removed.
-func (s *Store) SplitRange(
-	ctx context.Context, leftRepl, rightReplOrNil *Replica, split *roachpb.SplitTrigger,
-) error {
-	rightDesc := &split.RightDesc
-	newLeftDesc := &split.LeftDesc
-	oldLeftDesc := leftRepl.Desc()
-	if !bytes.Equal(oldLeftDesc.EndKey, rightDesc.EndKey) ||
-		bytes.Compare(oldLeftDesc.StartKey, rightDesc.StartKey) >= 0 {
-		return errors.Errorf("left range is not splittable by right range: %+v, %+v", oldLeftDesc, rightDesc)
-	}
-
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if exRng, ok := s.mu.uninitReplicas[rightDesc.RangeID]; rightReplOrNil != nil && ok {
-		// If we have an uninitialized replica of the new range we require pointer
-		// equivalence with rightRepl. See Store.splitTriggerPostApply().
-		if exRng != rightReplOrNil {
-			log.Fatalf(ctx, "found unexpected uninitialized replica: %s vs %s", exRng, rightReplOrNil)
-		}
-		// NB: We only remove from uninitReplicas and the replicaQueues maps here
-		// so that we don't leave open a window where a replica is temporarily not
-		// present in Store.mu.replicas.
-		delete(s.mu.uninitReplicas, rightDesc.RangeID)
-		s.replicaQueues.Delete(int64(rightDesc.RangeID))
-	}
-
-	leftRepl.setDesc(ctx, newLeftDesc)
-
-	// Clear the LHS txn wait queue, to redirect to the RHS if
-	// appropriate. We do this after setDescWithoutProcessUpdate
-	// to ensure that no pre-split commands are inserted into the
-	// txnWaitQueue after we clear it.
-	leftRepl.txnWaitQueue.Clear(false /* disable */)
-
-	// The rangefeed processor will no longer be provided logical ops for
-	// its entire range, so it needs to be shut down and all registrations
-	// need to retry.
-	// TODO(nvanbenschoten): It should be possible to only reject registrations
-	// that overlap with the new range of the split and keep registrations that
-	// are only interested in keys that are still on the original range running.
-	leftRepl.disconnectRangefeedWithReason(
-		roachpb.RangeFeedRetryError_REASON_RANGE_SPLIT,
-	)
-
-	// Clear the original range's request stats, since they include requests for
-	// spans that are now owned by the new range.
-	leftRepl.leaseholderStats.resetRequestCounts()
-
-	if rightReplOrNil == nil {
-		throwawayRightWriteStats := new(replicaStats)
-		leftRepl.writeStats.splitRequestCounts(throwawayRightWriteStats)
-	} else {
-		rightRepl := rightReplOrNil
-		leftRepl.writeStats.splitRequestCounts(rightRepl.writeStats)
-		if err := s.addReplicaInternalLocked(rightRepl); err != nil {
-			return errors.Errorf("unable to add replica %v: %s", rightRepl, err)
-		}
-
-		// Update the replica's cached byte thresholds. This is a no-op if the system
-		// config is not available, in which case we rely on the next gossip update
-		// to perform the update.
-		if err := rightRepl.updateRangeInfo(rightRepl.Desc()); err != nil {
-			return err
-		}
-		// Add the range to metrics and maybe gossip on capacity change.
-		s.metrics.ReplicaCount.Inc(1)
-		s.maybeGossipOnCapacityChange(ctx, rangeAddEvent)
-	}
-
-	return nil
-}
-
-// MergeRange expands the left-hand replica, leftRepl, to absorb the right-hand
-// replica, identified by rightDesc. freezeStart specifies the time at which the
-// right-hand replica promised to stop serving traffic and is used to initialize
-// the timestamp cache's low water mark for the right-hand keyspace. The
-// right-hand replica must exist on this store and the raftMus for both the
-// left-hand and right-hand replicas must be held.
-func (s *Store) MergeRange(
-	ctx context.Context,
-	leftRepl *Replica,
-	newLeftDesc, rightDesc roachpb.RangeDescriptor,
-	freezeStart hlc.Timestamp,
-) error {
-	if oldLeftDesc := leftRepl.Desc(); !oldLeftDesc.EndKey.Less(newLeftDesc.EndKey) {
-		return errors.Errorf("the new end key is not greater than the current one: %+v <= %+v",
-			newLeftDesc.EndKey, oldLeftDesc.EndKey)
-	}
-
-	rightRepl, err := s.GetReplica(rightDesc.RangeID)
-	if err != nil {
-		return err
-	}
-
-	leftRepl.raftMu.AssertHeld()
-	rightRepl.raftMu.AssertHeld()
-
-	// Shut down rangefeed processors on either side of the merge.
-	//
-	// It isn't strictly necessary to shut-down a rangefeed processor on the
-	// surviving replica in a merge, but we choose to in order to avoid clients
-	// who were monitoring both sides of the merge from establishing multiple
-	// partial rangefeeds to the surviving range.
-	// TODO(nvanbenschoten): does this make sense? We could just adjust the
-	// bounds of the leftRepl.Processor.
-	//
-	// NB: removeInitializedReplicaRaftMuLocked also disconnects any initialized
-	// rangefeeds with REASON_REPLICA_REMOVED. That's ok because we will have
-	// already disconnected the rangefeed here.
-	leftRepl.disconnectRangefeedWithReason(
-		roachpb.RangeFeedRetryError_REASON_RANGE_MERGED,
-	)
-	rightRepl.disconnectRangefeedWithReason(
-		roachpb.RangeFeedRetryError_REASON_RANGE_MERGED,
-	)
-
-	if err := rightRepl.postDestroyRaftMuLocked(ctx, rightRepl.GetMVCCStats()); err != nil {
-		return err
-	}
-
-	// Note that we were called (indirectly) from raft processing so we must
-	// call removeInitializedReplicaRaftMuLocked directly to avoid deadlocking
-	// on the right-hand replica's raftMu.
-	if err := s.removeInitializedReplicaRaftMuLocked(ctx, rightRepl, rightDesc.NextReplicaID, RemoveOptions{
-		DestroyData: false, // the replica was destroyed when the merge commit applied
-	}); err != nil {
-		return errors.Errorf("cannot remove range: %s", err)
-	}
-
-	if leftRepl.leaseholderStats != nil {
-		leftRepl.leaseholderStats.resetRequestCounts()
-	}
-	if leftRepl.writeStats != nil {
-		// Note: this could be drastically improved by adding a replicaStats method
-		// that merges stats. Resetting stats is typically bad for the rebalancing
-		// logic that depends on them.
-		leftRepl.writeStats.resetRequestCounts()
-	}
-
-	// Clear the wait queue to redirect the queued transactions to the
-	// left-hand replica, if necessary.
-	rightRepl.txnWaitQueue.Clear(true /* disable */)
-
-	leftLease, _ := leftRepl.GetLease()
-	rightLease, _ := rightRepl.GetLease()
-	if leftLease.OwnedBy(s.Ident.StoreID) && !rightLease.OwnedBy(s.Ident.StoreID) {
-		// We hold the lease for the LHS, but do not hold the lease for the RHS.
-		// That means we don't have up-to-date timestamp cache entries for the
-		// keyspace previously owned by the RHS. Bump the low water mark for the RHS
-		// keyspace to freezeStart, the time at which the RHS promised to stop
-		// serving traffic, as freezeStart is guaranteed to be greater than any
-		// entry in the RHS's timestamp cache.
-		//
-		// Note that we need to update our clock with freezeStart to preserve the
-		// invariant that our clock is always greater than or equal to any
-		// timestamps in the timestamp cache. For a full discussion, see the comment
-		// on TestStoreRangeMergeTimestampCacheCausality.
-		_ = s.Clock().Update(freezeStart)
-		setTimestampCacheLowWaterMark(s.tsCache, &rightDesc, freezeStart)
-	}
-
-	// Update the subsuming range's descriptor.
-	leftRepl.setDesc(ctx, &newLeftDesc)
-	return nil
-}
-
-// addReplicaInternalLocked adds the replica to the replicas map and the
-// replicasByKey btree. Returns an error if a replica with
-// the same Range ID or a KeyRange that overlaps has already been added to
-// this store. addReplicaInternalLocked requires that the store lock is held.
-func (s *Store) addReplicaInternalLocked(repl *Replica) error {
-	if !repl.IsInitialized() {
-		return errors.Errorf("attempted to add uninitialized replica %s", repl)
-	}
-
-	if err := s.addReplicaToRangeMapLocked(repl); err != nil {
-		return err
-	}
-
-	if exRange := s.getOverlappingKeyRangeLocked(repl.Desc()); exRange != nil {
-		return errors.Errorf("%s: cannot addReplicaInternalLocked; range %s has overlapping range %s", s, repl, exRange.Desc())
-	}
-
-	if exRngItem := s.mu.replicasByKey.ReplaceOrInsert(repl); exRngItem != nil {
-		return errors.Errorf("%s: cannot addReplicaInternalLocked; range for key %v already exists in replicasByKey btree", s,
-			exRngItem.(KeyRange).startKey())
-	}
-
-	return nil
-}
-
-// addPlaceholderLocked adds the specified placeholder. Requires that the
-// raftMu of the replica whose place is being held is locked.
-func (s *Store) addPlaceholder(placeholder *ReplicaPlaceholder) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.addPlaceholderLocked(placeholder)
-}
-
-// addPlaceholderLocked adds the specified placeholder. Requires that Store.mu
-// and the raftMu of the replica whose place is being held are locked.
-func (s *Store) addPlaceholderLocked(placeholder *ReplicaPlaceholder) error {
-	rangeID := placeholder.Desc().RangeID
-	if exRng := s.mu.replicasByKey.ReplaceOrInsert(placeholder); exRng != nil {
-		return errors.Errorf("%s overlaps with existing KeyRange %s in replicasByKey btree", placeholder, exRng)
-	}
-	if exRng, ok := s.mu.replicaPlaceholders[rangeID]; ok {
-		return errors.Errorf("%s has ID collision with existing KeyRange %s", placeholder, exRng)
-	}
-	s.mu.replicaPlaceholders[rangeID] = placeholder
-	return nil
-}
-
-// removePlaceholder removes a placeholder for the specified range if it
-// exists, returning true if a placeholder was present and removed and false
-// otherwise. Requires that the raftMu of the replica whose place is being held
-// is locked.
-func (s *Store) removePlaceholder(ctx context.Context, rngID roachpb.RangeID) bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.removePlaceholderLocked(ctx, rngID)
-}
-
-// removePlaceholderLocked removes the specified placeholder. Requires that
-// Store.mu and the raftMu of the replica whose place is being held are locked.
-func (s *Store) removePlaceholderLocked(ctx context.Context, rngID roachpb.RangeID) bool {
-	placeholder, ok := s.mu.replicaPlaceholders[rngID]
-	if !ok {
-		return false
-	}
-	switch exRng := s.mu.replicasByKey.Delete(placeholder).(type) {
-	case *ReplicaPlaceholder:
-		delete(s.mu.replicaPlaceholders, rngID)
-		if exRng2 := s.getOverlappingKeyRangeLocked(&exRng.rangeDesc); exRng2 != nil {
-			log.Fatalf(ctx, "corrupted replicasByKey map: %s and %s overlapped", exRng, exRng2)
-		}
-		return true
-	case nil:
-		log.Fatalf(ctx, "r%d: placeholder not found", rngID)
-	default:
-		log.Fatalf(ctx, "r%d: expected placeholder, got %T", rngID, exRng)
-	}
-	return false // appease the compiler
-}
-
-// addReplicaToRangeMapLocked adds the replica to the replicas map.
-func (s *Store) addReplicaToRangeMapLocked(repl *Replica) error {
-	// It's ok for the replica to exist in the replicas map as long as it is the
-	// same replica object. This occurs during splits where the right-hand side
-	// is added to the replicas map before it is initialized.
-	if existing, loaded := s.mu.replicas.LoadOrStore(
-		int64(repl.RangeID), unsafe.Pointer(repl)); loaded && (*Replica)(existing) != repl {
-		return errors.Errorf("%s: replica already exists", repl)
-	}
-	// Check whether the replica is unquiesced but not in the map. This
-	// can happen during splits and merges, where the uninitialized (but
-	// also unquiesced) replica is removed from the unquiesced replica
-	// map in advance of this method being called.
-	s.unquiescedReplicas.Lock()
-	if _, ok := s.unquiescedReplicas.m[repl.RangeID]; !repl.mu.quiescent && !ok {
-		s.unquiescedReplicas.m[repl.RangeID] = struct{}{}
-	}
-	s.unquiescedReplicas.Unlock()
-	return nil
-}
-
-// RemoveOptions bundles boolean parameters for Store.RemoveReplica.
-type RemoveOptions struct {
-	DestroyData bool
-
-	// ignoreDestroyStatus allows a caller to instruct the store to remove
-	// replicas which are already marked as destroyed. This is helpful in cases
-	// where the caller knows that it set the destroy status and cannot have raced
-	// with another goroutine. See Replica.handleChangeReplicasResult().
-	ignoreDestroyStatus bool
-}
-
-// RemoveReplica removes the replica from the store's replica map and from the
-// sorted replicasByKey btree.
-//
-// The NextReplicaID from the replica descriptor that was used to make the
-// removal decision is passed in. Removal is aborted if the replica ID has
-// advanced to or beyond the NextReplicaID since the removal decision was made.
-//
-// If opts.DestroyReplica is false, replica.destroyRaftMuLocked is not called.
-//
-// The passed replica must be initialized.
-func (s *Store) RemoveReplica(
-	ctx context.Context, rep *Replica, nextReplicaID roachpb.ReplicaID, opts RemoveOptions,
-) error {
-	rep.raftMu.Lock()
-	defer rep.raftMu.Unlock()
-	return s.removeInitializedReplicaRaftMuLocked(ctx, rep, nextReplicaID, opts)
-}
-
-// removeReplicaRaftMuLocked removes the passed replica. If the replica is
-// initialized the RemoveOptions will be consulted.
-func (s *Store) removeReplicaRaftMuLocked(
-	ctx context.Context, rep *Replica, nextReplicaID roachpb.ReplicaID, opts RemoveOptions,
-) error {
-	rep.raftMu.AssertHeld()
-	if rep.IsInitialized() {
-		return errors.Wrap(s.removeInitializedReplicaRaftMuLocked(ctx, rep, nextReplicaID, opts),
-			"failed to remove replica")
-	}
-	s.removeUninitializedReplicaRaftMuLocked(ctx, rep, nextReplicaID)
-	return nil
-}
-
-// removeInitializedReplicaRaftMuLocked is the implementation of RemoveReplica,
-// which is sometimes called directly when the necessary lock is already held.
-// It requires that Replica.raftMu is held and that s.mu is not held.
-func (s *Store) removeInitializedReplicaRaftMuLocked(
-	ctx context.Context, rep *Replica, nextReplicaID roachpb.ReplicaID, opts RemoveOptions,
-) error {
-	rep.raftMu.AssertHeld()
-
-	// Sanity checks before committing to the removal by setting the
-	// destroy status.
-	var desc *roachpb.RangeDescriptor
-	var replicaID roachpb.ReplicaID
-	{
-		rep.mu.Lock()
-
-		// Detect if we were already removed.
-		if !opts.ignoreDestroyStatus && rep.mu.destroyStatus.Removed() {
-			rep.mu.Unlock()
-			return nil // already removed, noop
-		}
-
-		// We check both rep.mu.ReplicaID and rep.mu.state.Desc's replica ID because
-		// they can differ in cases when a replica's ID is increased due to an
-		// incoming raft message (see #14231 for background).
-		replicaID = rep.mu.replicaID
-		if rep.mu.replicaID >= nextReplicaID {
-			rep.mu.Unlock()
-			return errors.Errorf("cannot remove replica %s; replica ID has changed (%s >= %s)",
-				rep, rep.mu.replicaID, nextReplicaID)
-		}
-		desc = rep.mu.state.Desc
-		if repDesc, ok := desc.GetReplicaDescriptor(s.StoreID()); ok && repDesc.ReplicaID >= nextReplicaID {
-			rep.mu.Unlock()
-			return errors.Errorf("cannot remove replica %s; replica descriptor's ID has changed (%s >= %s)",
-				rep, repDesc.ReplicaID, nextReplicaID)
-		}
-
-		// This is a fatal error as an initialized replica can never become
-		/// uninitialized.
-		if !rep.isInitializedRLocked() {
-			rep.mu.Unlock()
-			log.Fatalf(ctx, "uninitialized replica cannot be removed with removeInitializedReplica: %v",
-				rep)
-		}
-
-		// Mark the replica as removed before deleting data.
-		rep.mu.destroyStatus.Set(roachpb.NewRangeNotFoundError(rep.RangeID, rep.StoreID()),
-			destroyReasonRemoved)
-		rep.mu.Unlock()
-	}
-
-	// Proceed with the removal, all errors encountered from here down are fatal.
-
-	// Another sanity check that this replica is a part of this store.
-	existing, err := s.GetReplica(rep.RangeID)
-	if err != nil {
-		log.Fatalf(ctx, "cannot remove replica which does not exist: %v", err)
-	} else if existing != rep {
-		log.Fatalf(ctx, "replica %v replaced by %v before being removed",
-			rep, existing)
-	}
-
-	// During merges, the context might have the subsuming range, so we explicitly
-	// log the replica to be removed.
-	log.Infof(ctx, "removing replica r%d/%d", rep.RangeID, replicaID)
-
-	s.mu.Lock()
-	if placeholder := s.getOverlappingKeyRangeLocked(desc); placeholder != rep {
-		// This is a fatal error because uninitialized replicas shouldn't make it
-		// this far. This method will need some changes when we introduce GC of
-		// uninitialized replicas.
-		s.mu.Unlock()
-		log.Fatalf(ctx, "replica %+v unexpectedly overlapped by %+v", rep, placeholder)
-	}
-	// Adjust stats before calling Destroy. This can be called before or after
-	// Destroy, but this configuration helps avoid races in stat verification
-	// tests.
-	s.metrics.subtractMVCCStats(rep.GetMVCCStats())
-	s.metrics.ReplicaCount.Dec(1)
-	s.mu.Unlock()
-
-	// The replica will no longer exist, so cancel any rangefeed registrations.
-	rep.disconnectRangefeedWithReason(
-		roachpb.RangeFeedRetryError_REASON_REPLICA_REMOVED,
-	)
-
-	// Mark the replica as destroyed and (optionally) destroy the on-disk data
-	// while not holding Store.mu. This is safe because we're holding
-	// Replica.raftMu and the replica is present in Store.mu.replicasByKey
-	// (preventing any concurrent access to the replica's key range).
-
-	rep.readOnlyCmdMu.Lock()
-	rep.mu.Lock()
-	rep.cancelPendingCommandsLocked()
-	rep.mu.internalRaftGroup = nil
-	rep.mu.Unlock()
-	rep.readOnlyCmdMu.Unlock()
-
-	if opts.DestroyData {
-		if err := rep.destroyRaftMuLocked(ctx, nextReplicaID); err != nil {
-			return err
-		}
-	}
-
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.unlinkReplicaByRangeIDLocked(rep.RangeID)
-	if placeholder := s.mu.replicasByKey.Delete(rep); placeholder != rep {
-		// We already checked that our replica was present in replicasByKey
-		// above. Nothing should have been able to change that.
-		log.Fatalf(ctx, "replica %+v unexpectedly overlapped by %+v", rep, placeholder)
-	}
-	if rep2 := s.getOverlappingKeyRangeLocked(desc); rep2 != nil {
-		log.Fatalf(ctx, "corrupted replicasByKey map: %s and %s overlapped", rep, rep2)
-	}
-	delete(s.mu.replicaPlaceholders, rep.RangeID)
-	// TODO(peter): Could release s.mu.Lock() here.
-	s.maybeGossipOnCapacityChange(ctx, rangeRemoveEvent)
-	s.scanner.RemoveReplica(rep)
-	return nil
-}
-
-// removeUninitializedReplicaRaftMuLocked removes an uninitialized replica.
-// All paths which call this code held the raftMu before calling this method
-// and ensured that the removal was sane given the current replicaID and
-// initialization status (which only changes under the raftMu).
-func (s *Store) removeUninitializedReplicaRaftMuLocked(
-	ctx context.Context, rep *Replica, nextReplicaID roachpb.ReplicaID,
-) {
-	rep.raftMu.AssertHeld()
-
-	// Sanity check this removal and set the destroyStatus.
-	{
-		rep.mu.Lock()
-
-		// Detect if we were already removed, this is a fatal error
-		// because we should have already checked this under the raftMu
-		// before calling this method.
-		if rep.mu.destroyStatus.Removed() {
-			rep.mu.Unlock()
-			log.Fatalf(ctx, "uninitialized replica unexpectedly already removed")
-		}
-
-		if rep.isInitializedRLocked() {
-			rep.mu.Unlock()
-			log.Fatalf(ctx, "cannot remove initialized replica in removeUninitializedReplica: %v", rep)
-		}
-
-		// Mark the replica as removed before deleting data.
-		rep.mu.destroyStatus.Set(roachpb.NewRangeNotFoundError(rep.RangeID, rep.StoreID()),
-			destroyReasonRemoved)
-
-		rep.mu.Unlock()
-	}
-
-	// Proceed with the removal.
-
-	rep.readOnlyCmdMu.Lock()
-	rep.mu.Lock()
-	rep.cancelPendingCommandsLocked()
-	rep.mu.internalRaftGroup = nil
-	rep.mu.Unlock()
-	rep.readOnlyCmdMu.Unlock()
-
-	if err := rep.destroyRaftMuLocked(ctx, nextReplicaID); err != nil {
-		log.Fatalf(ctx, "failed to remove uninitialized replica %v: %v", rep, err)
-	}
-
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	// Sanity check, could be removed.
-	value, stillExists := s.mu.replicas.Load(int64(rep.RangeID))
-	if !stillExists {
-		log.Fatalf(ctx, "uninitialized replica was removed in the meantime")
-	}
-	existing := (*Replica)(value)
-	if existing == rep {
-		log.Infof(ctx, "removing uninitialized replica %v", rep)
-	} else {
-		log.Fatalf(ctx, "uninitialized replica %v was unexpectedly replaced", existing)
-	}
-
-	s.metrics.ReplicaCount.Dec(1)
-
-	// Only an uninitialized replica can have a placeholder since, by
-	// definition, an initialized replica will be present in the
-	// replicasByKey map. While the replica will usually consume the
-	// placeholder itself, that isn't guaranteed and so this invocation
-	// here is crucial (i.e. don't remove it).
-	if s.removePlaceholderLocked(ctx, rep.RangeID) {
-		atomic.AddInt32(&s.counts.droppedPlaceholders, 1)
-	}
-	s.unlinkReplicaByRangeIDLocked(rep.RangeID)
-}
-
-// unlinkReplicaByRangeIDLocked removes all of the store's references to the
-// provided replica that are keyed by its range ID. The replica may also need
-// to be removed from the replicasByKey map.
-//
-// store.mu must be held.
-func (s *Store) unlinkReplicaByRangeIDLocked(rangeID roachpb.RangeID) {
-	s.mu.AssertHeld()
-	s.unquiescedReplicas.Lock()
-	delete(s.unquiescedReplicas.m, rangeID)
-	s.unquiescedReplicas.Unlock()
-	delete(s.mu.uninitReplicas, rangeID)
-	s.replicaQueues.Delete(int64(rangeID))
-	s.mu.replicas.Delete(int64(rangeID))
-}
-
-// maybeMarkReplicaInitializedLocked should be called whenever a previously
-// unintialized replica has become initialized so that the store can update its
-// internal bookkeeping. It requires that Store.mu and Replica.raftMu
-// are locked.
-func (s *Store) maybeMarkReplicaInitializedLocked(ctx context.Context, repl *Replica) error {
-	if !repl.IsInitialized() {
-		return errors.Errorf("attempted to process uninitialized range %s", repl)
-	}
-
-	rangeID := repl.RangeID
-
-	if _, ok := s.mu.uninitReplicas[rangeID]; !ok {
-		// Do nothing if the range has already been initialized.
-		return nil
-	}
-	delete(s.mu.uninitReplicas, rangeID)
-
-	if exRange := s.getOverlappingKeyRangeLocked(repl.Desc()); exRange != nil {
-		return errors.Errorf("%s: cannot initialize replica; range %s has overlapping range %s",
-			s, repl, exRange.Desc())
-	}
-	if exRngItem := s.mu.replicasByKey.ReplaceOrInsert(repl); exRngItem != nil {
-		return errors.Errorf("range for key %v already exists in replicasByKey btree",
-			(exRngItem.(*Replica)).startKey())
-	}
-
-	// Add the range to metrics and maybe gossip on capacity change.
-	s.metrics.ReplicaCount.Inc(1)
-	s.maybeGossipOnCapacityChange(ctx, rangeAddEvent)
-
-	return nil
-}
-
 // Attrs returns the attributes of the underlying store.
 func (s *Store) Attrs() roachpb.Attributes {
 	return s.engine.Attrs()
@@ -2955,330 +2187,6 @@ func (s *Store) Descriptor(useCached bool) (*roachpb.StoreDescriptor, error) {
 	}, nil
 }
 
-// Send fetches a range based on the header's replica, assembles method, args &
-// reply into a Raft Cmd struct and executes the command using the fetched
-// range.
-//
-// An incoming request may be transactional or not. If it is not transactional,
-// the timestamp at which it executes may be higher than that optionally
-// specified through the incoming BatchRequest, and it is not guaranteed that
-// all operations are written at the same timestamp. If it is transactional, a
-// timestamp must not be set - it is deduced automatically from the
-// transaction. In particular, the read (original) timestamp will be used for
-// all reads and the write (provisional commit) timestamp will be used for
-// all writes. See the comments on txn.TxnMeta.Timestamp and txn.OrigTimestamp
-// for more details.
-//
-// Should a transactional operation be forced to a higher timestamp (for
-// instance due to the timestamp cache or finding a committed value in the path
-// of one of its writes), the response will have a transaction set which should
-// be used to update the client transaction object.
-func (s *Store) Send(
-	ctx context.Context, ba roachpb.BatchRequest,
-) (br *roachpb.BatchResponse, pErr *roachpb.Error) {
-	// Attach any log tags from the store to the context (which normally
-	// comes from gRPC).
-	ctx = s.AnnotateCtx(ctx)
-	for _, union := range ba.Requests {
-		arg := union.GetInner()
-		header := arg.Header()
-		if err := verifyKeys(header.Key, header.EndKey, roachpb.IsRange(arg)); err != nil {
-			return nil, roachpb.NewError(err)
-		}
-	}
-
-	// Limit the number of concurrent AddSSTable requests, since they're expensive
-	// and block all other writes to the same span.
-	if ba.IsSingleAddSSTableRequest() {
-		if err := s.limiters.ConcurrentAddSSTableRequests.Begin(ctx); err != nil {
-			return nil, roachpb.NewError(err)
-		}
-		defer s.limiters.ConcurrentAddSSTableRequests.Finish()
-
-		if err := s.limiters.AddSSTableRequestRate.Wait(ctx); err != nil {
-			return nil, roachpb.NewError(err)
-		}
-		s.engine.PreIngestDelay(ctx)
-	}
-
-	if err := ba.SetActiveTimestamp(s.Clock().Now); err != nil {
-		return nil, roachpb.NewError(err)
-	}
-
-	if s.cfg.TestingKnobs.ClockBeforeSend != nil {
-		s.cfg.TestingKnobs.ClockBeforeSend(s.cfg.Clock, ba)
-	}
-
-	// Update our clock with the incoming request timestamp. This advances the
-	// local node's clock to a high water mark from all nodes with which it has
-	// interacted. We hold on to the resulting timestamp - we know that any
-	// write with a higher timestamp we run into later must have started after
-	// this point in (absolute) time.
-	var now hlc.Timestamp
-	if s.cfg.TestingKnobs.DisableMaxOffsetCheck {
-		now = s.cfg.Clock.Update(ba.Timestamp)
-	} else {
-		// If the command appears to come from a node with a bad clock,
-		// reject it now before we reach that point.
-		var err error
-		if now, err = s.cfg.Clock.UpdateAndCheckMaxOffset(ba.Timestamp); err != nil {
-			return nil, roachpb.NewError(err)
-		}
-	}
-
-	defer func() {
-		if r := recover(); r != nil {
-			// On panic, don't run the defer. It's probably just going to panic
-			// again due to undefined state.
-			panic(r)
-		}
-		if ba.Txn != nil {
-			// We're in a Txn, so we can reduce uncertainty restarts by attaching
-			// the above timestamp to the returned response or error. The caller
-			// can use it to shorten its uncertainty interval when it comes back to
-			// this node.
-			if pErr != nil {
-				pErr.OriginNode = ba.Replica.NodeID
-				if txn := pErr.GetTxn(); txn == nil {
-					pErr.SetTxn(ba.Txn)
-				}
-			} else {
-				if br.Txn == nil {
-					br.Txn = ba.Txn
-				}
-				// Update our clock with the outgoing response txn timestamp
-				// (if timestamp has been forwarded).
-				if ba.Timestamp.Less(br.Txn.Timestamp) {
-					s.cfg.Clock.Update(br.Txn.Timestamp)
-				}
-			}
-		} else {
-			if pErr == nil {
-				// Update our clock with the outgoing response timestamp.
-				// (if timestamp has been forwarded).
-				if ba.Timestamp.Less(br.Timestamp) {
-					s.cfg.Clock.Update(br.Timestamp)
-				}
-			}
-		}
-
-		if pErr != nil {
-			pErr.Now = now
-		} else {
-			br.Now = now
-		}
-	}()
-
-	if ba.Txn != nil {
-		// We make our transaction aware that no other operation that causally
-		// precedes it could have started after `now`. This is important: If we
-		// wind up pushing a value, it will be in our immediate future, and not
-		// updating the top end of our uncertainty timestamp would lead to a
-		// restart (at least in the absence of a prior observed timestamp from
-		// this node, in which case the following is a no-op).
-		if _, ok := ba.Txn.GetObservedTimestamp(ba.Replica.NodeID); !ok {
-			txnClone := ba.Txn.Clone()
-			txnClone.UpdateObservedTimestamp(ba.Replica.NodeID, now)
-			ba.Txn = txnClone
-		}
-	}
-
-	if log.V(1) {
-		log.Eventf(ctx, "executing %s", ba)
-	} else if log.HasSpanOrEvent(ctx) {
-		log.Eventf(ctx, "executing %d requests", len(ba.Requests))
-	}
-
-	var cleanupAfterWriteIntentError func(newWIErr *roachpb.WriteIntentError, newIntentTxn *enginepb.TxnMeta)
-	defer func() {
-		if cleanupAfterWriteIntentError != nil {
-			// This request wrote an intent only if there was no error, the request
-			// is transactional, the transaction is not yet finalized, and the request
-			// wasn't read-only.
-			if pErr == nil && ba.Txn != nil && !br.Txn.Status.IsFinalized() && !ba.IsReadOnly() {
-				cleanupAfterWriteIntentError(nil, &br.Txn.TxnMeta)
-			} else {
-				cleanupAfterWriteIntentError(nil, nil)
-			}
-		}
-	}()
-
-	// Add the command to the range for execution; exit retry loop on success.
-	for {
-		// Exit loop if context has been canceled or timed out.
-		if err := ctx.Err(); err != nil {
-			return nil, roachpb.NewError(errors.Wrap(err, "aborted during Store.Send"))
-		}
-
-		// Get range and add command to the range for execution.
-		repl, err := s.GetReplica(ba.RangeID)
-		if err != nil {
-			return nil, roachpb.NewError(err)
-		}
-		if !repl.IsInitialized() {
-			repl.mu.RLock()
-			replicaID := repl.mu.replicaID
-			repl.mu.RUnlock()
-
-			// If we have an uninitialized copy of the range, then we are
-			// probably a valid member of the range, we're just in the
-			// process of getting our snapshot. If we returned
-			// RangeNotFoundError, the client would invalidate its cache,
-			// but we can be smarter: the replica that caused our
-			// uninitialized replica to be created is most likely the
-			// leader.
-			return nil, roachpb.NewError(&roachpb.NotLeaseHolderError{
-				RangeID:     ba.RangeID,
-				LeaseHolder: repl.creatingReplica,
-				// The replica doesn't have a range descriptor yet, so we have to build
-				// a ReplicaDescriptor manually.
-				Replica: roachpb.ReplicaDescriptor{
-					NodeID:    repl.store.nodeDesc.NodeID,
-					StoreID:   repl.store.StoreID(),
-					ReplicaID: replicaID,
-				},
-			})
-		}
-
-		// If necessary, the request may need to wait in the txn wait queue,
-		// pending updates to the target transaction for either PushTxn or
-		// QueryTxn requests.
-		if br, pErr = s.maybeWaitForPushee(ctx, &ba, repl); br != nil || pErr != nil {
-			return br, pErr
-		}
-		br, pErr = repl.Send(ctx, ba)
-		if pErr == nil {
-			return br, nil
-		}
-
-		// Handle push txn failures and write intent conflicts locally and
-		// retry. Other errors are returned to caller.
-		switch t := pErr.GetDetail().(type) {
-		case *roachpb.TransactionPushError:
-			// On a transaction push error, retry immediately if doing so will
-			// enqueue into the txnWaitQueue in order to await further updates to
-			// the unpushed txn's status. We check ShouldPushImmediately to avoid
-			// retrying non-queueable PushTxnRequests (see #18191).
-			dontRetry := s.cfg.TestingKnobs.DontRetryPushTxnFailures
-			if !dontRetry && ba.IsSinglePushTxnRequest() {
-				pushReq := ba.Requests[0].GetInner().(*roachpb.PushTxnRequest)
-				dontRetry = txnwait.ShouldPushImmediately(pushReq)
-			}
-			if dontRetry {
-				// If we're not retrying on push txn failures return a txn retry error
-				// after the first failure to guarantee a retry.
-				if ba.Txn != nil {
-					err := roachpb.NewTransactionRetryError(
-						roachpb.RETRY_REASON_UNKNOWN, "DontRetryPushTxnFailures testing knob")
-					return nil, roachpb.NewErrorWithTxn(err, ba.Txn)
-				}
-				return nil, pErr
-			}
-
-			// Enqueue unsuccessfully pushed transaction on the txnWaitQueue and
-			// retry the command.
-			repl.txnWaitQueue.Enqueue(&t.PusheeTxn)
-			pErr = nil
-
-		case *roachpb.IndeterminateCommitError:
-			if s.cfg.TestingKnobs.DontRecoverIndeterminateCommits {
-				return nil, pErr
-			}
-			// On an indeterminate commit error, attempt to recover and finalize
-			// the stuck transaction. Retry immediately if successful.
-			if _, err := s.recoveryMgr.ResolveIndeterminateCommit(ctx, t); err != nil {
-				// Do not propagate ambiguous results; assume success and retry original op.
-				if _, ok := err.(*roachpb.AmbiguousResultError); !ok {
-					// Preserve the error index.
-					index := pErr.Index
-					pErr = roachpb.NewError(err)
-					pErr.Index = index
-					return nil, pErr
-				}
-			}
-			// We've recovered the transaction that blocked the push; retry command.
-			pErr = nil
-
-		case *roachpb.WriteIntentError:
-			// Process and resolve write intent error. We do this here because
-			// this is the code path with the requesting client waiting.
-			if pErr.Index != nil {
-				var pushType roachpb.PushTxnType
-				if ba.IsWrite() {
-					pushType = roachpb.PUSH_ABORT
-				} else {
-					pushType = roachpb.PUSH_TIMESTAMP
-				}
-
-				index := pErr.Index
-				args := ba.Requests[index.Index].GetInner()
-				// Make a copy of the header for the upcoming push; we will update
-				// the timestamp.
-				h := ba.Header
-				if h.Txn != nil {
-					// We must push at least to h.Timestamp, but in fact we want to
-					// go all the way up to a timestamp which was taken off the HLC
-					// after our operation started. This allows us to not have to
-					// restart for uncertainty as we come back and read.
-					obsTS, ok := h.Txn.GetObservedTimestamp(ba.Replica.NodeID)
-					if !ok {
-						// This was set earlier in this method, so it's
-						// completely unexpected to not be found now.
-						log.Fatalf(ctx, "missing observed timestamp: %+v", h.Txn)
-					}
-					h.Timestamp.Forward(obsTS)
-					// We are going to hand the header (and thus the transaction proto)
-					// to the RPC framework, after which it must not be changed (since
-					// that could race). Since the subsequent execution of the original
-					// request might mutate the transaction, make a copy here.
-					//
-					// See #9130.
-					h.Txn = h.Txn.Clone()
-				}
-				// Handle the case where we get more than one write intent error;
-				// we need to cleanup the previous attempt to handle it to allow
-				// any other pusher queued up behind this RPC to proceed.
-				if cleanupAfterWriteIntentError != nil {
-					cleanupAfterWriteIntentError(t, nil)
-				}
-				if cleanupAfterWriteIntentError, pErr =
-					s.intentResolver.ProcessWriteIntentError(ctx, pErr, args, h, pushType); pErr != nil {
-					// Do not propagate ambiguous results; assume success and retry original op.
-					if _, ok := pErr.GetDetail().(*roachpb.AmbiguousResultError); !ok {
-						// Preserve the error index.
-						pErr.Index = index
-						return nil, pErr
-					}
-					pErr = nil
-				}
-				// We've resolved the write intent; retry command.
-			}
-
-		case *roachpb.MergeInProgressError:
-			// A merge was in progress. We need to retry the command after the merge
-			// completes, as signaled by the closing of the replica's mergeComplete
-			// channel. Note that the merge may have already completed, in which case
-			// its mergeComplete channel will be nil.
-			mergeCompleteCh := repl.getMergeCompleteCh()
-			if mergeCompleteCh != nil {
-				select {
-				case <-mergeCompleteCh:
-					// Merge complete. Retry the command.
-				case <-ctx.Done():
-					return nil, roachpb.NewError(errors.Wrap(ctx.Err(), "aborted during merge"))
-				case <-s.stopper.ShouldQuiesce():
-					return nil, roachpb.NewError(&roachpb.NodeUnavailableError{})
-				}
-			}
-			pErr = nil
-		}
-
-		if pErr != nil {
-			return nil, pErr
-		}
-	}
-}
-
 // RangeFeed registers a rangefeed over the specified span. It sends updates to
 // the provided stream and returns with an optional error when the rangefeed is
 // complete.
@@ -3303,1004 +2211,6 @@ func (s *Store) RangeFeed(
 		return roachpb.NewError(roachpb.NewRangeNotFoundError(args.RangeID, s.StoreID()))
 	}
 	return repl.RangeFeed(args, stream)
-}
-
-// maybeWaitForPushee potentially diverts the incoming request to
-// the txnwait.Queue, where it will wait for updates to the target
-// transaction.
-func (s *Store) maybeWaitForPushee(
-	ctx context.Context, ba *roachpb.BatchRequest, repl *Replica,
-) (*roachpb.BatchResponse, *roachpb.Error) {
-	// If this is a push txn request, check the push queue first, which
-	// may cause this request to wait and either return a successful push
-	// txn response or else allow this request to proceed.
-	if ba.IsSinglePushTxnRequest() {
-		pushReq := ba.Requests[0].GetInner().(*roachpb.PushTxnRequest)
-		pushResp, pErr := repl.txnWaitQueue.MaybeWaitForPush(repl.AnnotateCtx(ctx), repl, pushReq)
-		// Copy the request in anticipation of setting the force arg and
-		// updating the Now timestamp (see below).
-		pushReqCopy := *pushReq
-		if pErr == txnwait.ErrDeadlock {
-			// We've experienced a deadlock; set Force=true on push request,
-			// and set the push type to ABORT.
-			pushReqCopy.Force = true
-			pushReqCopy.PushType = roachpb.PUSH_ABORT
-		} else if pErr != nil {
-			return nil, pErr
-		} else if pushResp != nil {
-			br := &roachpb.BatchResponse{}
-			br.Add(pushResp)
-			return br, nil
-		}
-		// Move the push timestamp forward to the current time, as this
-		// request may have been waiting to push the txn. If we don't
-		// move the timestamp forward to the current time, we may fail
-		// to push a txn which has expired.
-		now := s.Clock().Now()
-		ba.Timestamp.Forward(now)
-		ba.Requests = nil
-		ba.Add(&pushReqCopy)
-	} else if ba.IsSingleQueryTxnRequest() {
-		// For query txn requests, wait in the txn wait queue either for
-		// transaction update or for dependent transactions to change.
-		queryReq := ba.Requests[0].GetInner().(*roachpb.QueryTxnRequest)
-		pErr := repl.txnWaitQueue.MaybeWaitForQuery(repl.AnnotateCtx(ctx), repl, queryReq)
-		if pErr != nil {
-			return nil, pErr
-		}
-	}
-
-	return nil, nil
-}
-
-// HandleSnapshot reads an incoming streaming snapshot and applies it if
-// possible.
-func (s *Store) HandleSnapshot(
-	header *SnapshotRequest_Header, stream SnapshotResponseStream,
-) error {
-	ctx := s.AnnotateCtx(stream.Context())
-	const name = "storage.Store: handle snapshot"
-	return s.stopper.RunTaskWithErr(ctx, name, func(ctx context.Context) error {
-		s.metrics.raftRcvdMessages[raftpb.MsgSnap].Inc(1)
-
-		if s.IsDraining() {
-			return stream.Send(&SnapshotResponse{
-				Status:  SnapshotResponse_DECLINED,
-				Message: storeDrainingMsg,
-			})
-		}
-
-		return s.receiveSnapshot(ctx, header, stream)
-	})
-}
-
-// learnerType exists to avoid allocating on every coalesced beat to a learner.
-var learnerType = roachpb.LEARNER
-
-func (s *Store) uncoalesceBeats(
-	ctx context.Context,
-	beats []RaftHeartbeat,
-	fromReplica, toReplica roachpb.ReplicaDescriptor,
-	msgT raftpb.MessageType,
-	respStream RaftMessageResponseStream,
-) {
-	if len(beats) == 0 {
-		return
-	}
-	if log.V(4) {
-		log.Infof(ctx, "uncoalescing %d beats of type %v: %+v", len(beats), msgT, beats)
-	}
-	beatReqs := make([]RaftMessageRequest, len(beats))
-	for i, beat := range beats {
-		msg := raftpb.Message{
-			Type:   msgT,
-			From:   uint64(beat.FromReplicaID),
-			To:     uint64(beat.ToReplicaID),
-			Term:   beat.Term,
-			Commit: beat.Commit,
-		}
-		beatReqs[i] = RaftMessageRequest{
-			RangeID: beat.RangeID,
-			FromReplica: roachpb.ReplicaDescriptor{
-				NodeID:    fromReplica.NodeID,
-				StoreID:   fromReplica.StoreID,
-				ReplicaID: beat.FromReplicaID,
-			},
-			ToReplica: roachpb.ReplicaDescriptor{
-				NodeID:    toReplica.NodeID,
-				StoreID:   toReplica.StoreID,
-				ReplicaID: beat.ToReplicaID,
-			},
-			Message: msg,
-			Quiesce: beat.Quiesce,
-		}
-		if beat.ToIsLearner {
-			beatReqs[i].ToReplica.Type = &learnerType
-		}
-		if log.V(4) {
-			log.Infof(ctx, "uncoalesced beat: %+v", beatReqs[i])
-		}
-
-		if err := s.HandleRaftUncoalescedRequest(ctx, &beatReqs[i], respStream); err != nil {
-			log.Errorf(ctx, "could not handle uncoalesced heartbeat %s", err)
-		}
-	}
-}
-
-// HandleRaftRequest dispatches a raft message to the appropriate Replica. It
-// requires that s.mu is not held.
-func (s *Store) HandleRaftRequest(
-	ctx context.Context, req *RaftMessageRequest, respStream RaftMessageResponseStream,
-) *roachpb.Error {
-	// NB: unlike the other two RaftMessageHandler methods implemented by Store,
-	// this one doesn't need to directly run through a Stopper task because it
-	// delegates all work through a raftScheduler, whose workers' lifetimes are
-	// already tied to the Store's Stopper.
-	if len(req.Heartbeats)+len(req.HeartbeatResps) > 0 {
-		if req.RangeID != 0 {
-			log.Fatalf(ctx, "coalesced heartbeats must have rangeID == 0")
-		}
-		s.uncoalesceBeats(ctx, req.Heartbeats, req.FromReplica, req.ToReplica, raftpb.MsgHeartbeat, respStream)
-		s.uncoalesceBeats(ctx, req.HeartbeatResps, req.FromReplica, req.ToReplica, raftpb.MsgHeartbeatResp, respStream)
-		return nil
-	}
-	return s.HandleRaftUncoalescedRequest(ctx, req, respStream)
-}
-
-// HandleRaftUncoalescedRequest dispatches a raft message to the appropriate
-// Replica. It requires that s.mu is not held.
-func (s *Store) HandleRaftUncoalescedRequest(
-	ctx context.Context, req *RaftMessageRequest, respStream RaftMessageResponseStream,
-) *roachpb.Error {
-
-	if len(req.Heartbeats)+len(req.HeartbeatResps) > 0 {
-		log.Fatalf(ctx, "HandleRaftUncoalescedRequest cannot be given coalesced heartbeats or heartbeat responses, received %s", req)
-	}
-	// HandleRaftRequest is called on locally uncoalesced heartbeats (which are
-	// not sent over the network if the environment variable is set) so do not
-	// count them.
-	s.metrics.raftRcvdMessages[req.Message.Type].Inc(1)
-
-	value, ok := s.replicaQueues.Load(int64(req.RangeID))
-	if !ok {
-		value, _ = s.replicaQueues.LoadOrStore(int64(req.RangeID), unsafe.Pointer(&raftRequestQueue{}))
-	}
-	q := (*raftRequestQueue)(value)
-	q.Lock()
-	if len(q.infos) >= replicaRequestQueueSize {
-		q.Unlock()
-		// TODO(peter): Return an error indicating the request was dropped. Note
-		// that dropping the request is safe. Raft will retry.
-		s.metrics.RaftRcvdMsgDropped.Inc(1)
-		return nil
-	}
-	q.infos = append(q.infos, raftRequestInfo{
-		req:        req,
-		respStream: respStream,
-	})
-	first := len(q.infos) == 1
-	q.Unlock()
-
-	// processRequestQueue will process all infos in the slice each time it
-	// runs, so we only need to schedule a Raft request event if we added the
-	// first info in the slice. Everyone else can rely on the request that added
-	// the first info already having scheduled a Raft request event.
-	if first {
-		s.scheduler.EnqueueRaftRequest(req.RangeID)
-	}
-	return nil
-}
-
-// withReplicaForRequest calls the supplied function with the (lazily
-// initialized) Replica specified in the request. The replica passed to
-// the function will have its Replica.raftMu locked.
-func (s *Store) withReplicaForRequest(
-	ctx context.Context, req *RaftMessageRequest, f func(context.Context, *Replica) *roachpb.Error,
-) *roachpb.Error {
-	// Lazily create the replica.
-	r, _, err := s.getOrCreateReplica(
-		ctx,
-		req.RangeID,
-		req.ToReplica.ReplicaID,
-		&req.FromReplica,
-		req.ToReplica.GetType() == roachpb.LEARNER,
-	)
-	if err != nil {
-		return roachpb.NewError(err)
-	}
-	defer r.raftMu.Unlock()
-	ctx = r.AnnotateCtx(ctx)
-	r.setLastReplicaDescriptors(req)
-	return f(ctx, r)
-}
-
-// processRaftRequestWithReplica processes the (non-snapshot) Raft request on
-// the specified replica. Notably, it does not handle updates to the Raft Ready
-// state; callers will probably want to handle this themselves at some point.
-func (s *Store) processRaftRequestWithReplica(
-	ctx context.Context, r *Replica, req *RaftMessageRequest,
-) *roachpb.Error {
-	if verboseRaftLoggingEnabled() {
-		log.Infof(ctx, "incoming raft message:\n%s", raftDescribeMessage(req.Message, raftEntryFormatter))
-	}
-
-	if req.Message.Type == raftpb.MsgSnap {
-		log.Fatalf(ctx, "unexpected snapshot: %+v", req)
-	}
-
-	if req.Quiesce {
-		if req.Message.Type != raftpb.MsgHeartbeat {
-			log.Fatalf(ctx, "unexpected quiesce: %+v", req)
-		}
-		status := r.RaftStatus()
-		if status != nil && status.Term == req.Message.Term && status.Commit == req.Message.Commit {
-			if r.quiesce() {
-				return nil
-			}
-		}
-		if log.V(4) {
-			log.Infof(ctx, "not quiescing: local raft status is %+v, incoming quiesce message is %+v", status, req.Message)
-		}
-	}
-
-	if req.ToReplica.ReplicaID == 0 {
-		log.VEventf(ctx, 1, "refusing incoming Raft message %s from %+v to %+v",
-			req.Message.Type, req.FromReplica, req.ToReplica)
-		return roachpb.NewErrorf(
-			"cannot recreate replica that is not a member of its range (StoreID %s not found in r%d)",
-			r.store.StoreID(), req.RangeID,
-		)
-	}
-
-	drop := maybeDropMsgApp(ctx, (*replicaMsgAppDropper)(r), &req.Message, req.RangeStartKey)
-	if !drop {
-		if err := r.stepRaftGroup(req); err != nil {
-			return roachpb.NewError(err)
-		}
-	}
-	return nil
-}
-
-// processRaftSnapshotRequest processes the incoming non-preemptive snapshot
-// Raft request on the request's specified replica. The function makes sure to
-// handle any updated Raft Ready state. It also adds and later removes the
-// (potentially) necessary placeholder to protect against concurrent access to
-// the keyspace encompassed by the snapshot but not yet guarded by the replica.
-func (s *Store) processRaftSnapshotRequest(
-	ctx context.Context, snapHeader *SnapshotRequest_Header, inSnap IncomingSnapshot,
-) *roachpb.Error {
-	if snapHeader.IsPreemptive() {
-		return roachpb.NewError(crdberrors.AssertionFailedf(`expected a raft or learner snapshot`))
-	}
-
-	return s.withReplicaForRequest(ctx, &snapHeader.RaftMessageRequest, func(
-		ctx context.Context, r *Replica,
-	) (pErr *roachpb.Error) {
-		if snapHeader.RaftMessageRequest.Message.Type != raftpb.MsgSnap {
-			log.Fatalf(ctx, "expected snapshot: %+v", snapHeader.RaftMessageRequest)
-		}
-
-		// Check to see if a snapshot can be applied. Snapshots can always be applied
-		// to initialized replicas. Note that if we add a placeholder we need to
-		// already be holding Replica.raftMu in order to prevent concurrent
-		// raft-ready processing of uninitialized replicas.
-		var addedPlaceholder bool
-		var removePlaceholder bool
-		if err := func() error {
-			s.mu.Lock()
-			defer s.mu.Unlock()
-			placeholder, err := s.canApplySnapshotLocked(ctx, snapHeader)
-			if err != nil {
-				// If the storage cannot accept the snapshot, return an
-				// error before passing it to RawNode.Step, since our
-				// error handling options past that point are limited.
-				log.Infof(ctx, "cannot apply snapshot: %s", err)
-				return err
-			}
-
-			if placeholder != nil {
-				// NB: The placeholder added here is either removed below after a
-				// preemptive snapshot is applied or after the next call to
-				// Replica.handleRaftReady. Note that we can only get here if the
-				// replica doesn't exist or is uninitialized.
-				if err := s.addPlaceholderLocked(placeholder); err != nil {
-					log.Fatalf(ctx, "could not add vetted placeholder %s: %+v", placeholder, err)
-				}
-				addedPlaceholder = true
-			}
-			return nil
-		}(); err != nil {
-			return roachpb.NewError(err)
-		}
-
-		if addedPlaceholder {
-			// If we added a placeholder remove it before we return unless some other
-			// part of the code takes ownership of the removal (indicated by setting
-			// removePlaceholder to false).
-			removePlaceholder = true
-			defer func() {
-				if removePlaceholder {
-					if s.removePlaceholder(ctx, snapHeader.RaftMessageRequest.RangeID) {
-						atomic.AddInt32(&s.counts.removedPlaceholders, 1)
-					}
-				}
-			}()
-		}
-		// NB: we cannot get errRemoved here because we're promised by
-		// withReplicaForRequest that this replica is not currently being removed
-		// and we've been holding the raftMu the entire time.
-		if err := r.stepRaftGroup(&snapHeader.RaftMessageRequest); err != nil {
-			return roachpb.NewError(err)
-		}
-		_, expl, err := r.handleRaftReadyRaftMuLocked(ctx, inSnap)
-		maybeFatalOnRaftReadyErr(ctx, expl, err)
-		removePlaceholder = false
-		return nil
-	})
-}
-
-// HandleRaftResponse implements the RaftMessageHandler interface. Per the
-// interface specification, an error is returned if and only if the underlying
-// Raft connection should be closed.
-// It requires that s.mu is not held.
-func (s *Store) HandleRaftResponse(ctx context.Context, resp *RaftMessageResponse) error {
-	ctx = s.AnnotateCtx(ctx)
-	const name = "storage.Store: handle raft response"
-	return s.stopper.RunTaskWithErr(ctx, name, func(ctx context.Context) error {
-		repl, replErr := s.GetReplica(resp.RangeID)
-		if replErr == nil {
-			// Best-effort context annotation of replica.
-			ctx = repl.AnnotateCtx(ctx)
-		}
-		switch val := resp.Union.GetValue().(type) {
-		case *roachpb.Error:
-			switch tErr := val.GetDetail().(type) {
-			case *roachpb.ReplicaTooOldError:
-				if replErr != nil {
-					// RangeNotFoundErrors are expected here; nothing else is.
-					if _, ok := replErr.(*roachpb.RangeNotFoundError); !ok {
-						log.Error(ctx, replErr)
-					}
-					return nil
-				}
-
-				// Grab the raftMu in addition to the replica mu because
-				// cancelFailedProposalsLocked below requires it.
-				repl.raftMu.Lock()
-				defer repl.raftMu.Unlock()
-				repl.mu.Lock()
-
-				// If the replica ID in the error does not match then we know
-				// that the replica has been removed and re-added quickly. In
-				// that case, we don't want to add it to the replicaGCQueue.
-				// If the replica is not alive then we also should ignore this error.
-				if tErr.ReplicaID != repl.mu.replicaID ||
-					!repl.mu.destroyStatus.IsAlive() ||
-					// Ignore if we want to test the replicaGC queue.
-					s.TestingKnobs().DisableEagerReplicaRemoval {
-					repl.mu.Unlock()
-					return nil
-				}
-
-				// The replica will be garbage collected soon (we are sure
-				// since our replicaID is definitely too old), but in the meantime we
-				// already want to bounce all traffic from it. Note that the replica
-				// could be re-added with a higher replicaID, but we want to clear the
-				// replica's data before that happens.
-				if log.V(1) {
-					log.Infof(ctx, "setting local replica to destroyed due to ReplicaTooOld error")
-				}
-
-				repl.mu.Unlock()
-				nextReplicaID := tErr.ReplicaID + 1
-				return s.removeReplicaRaftMuLocked(ctx, repl, nextReplicaID, RemoveOptions{
-					DestroyData: true,
-				})
-			case *roachpb.RaftGroupDeletedError:
-				if replErr != nil {
-					// RangeNotFoundErrors are expected here; nothing else is.
-					if _, ok := replErr.(*roachpb.RangeNotFoundError); !ok {
-						log.Error(ctx, replErr)
-					}
-					return nil
-				}
-
-				// If the replica is talking to a replica that's been deleted, it must be
-				// out of date. While this may just mean it's slightly behind, it can
-				// also mean that it is so far behind it no longer knows where any of the
-				// other replicas are (#23994). Add it to the replica GC queue to do a
-				// proper check.
-				s.replicaGCQueue.AddAsync(ctx, repl, replicaGCPriorityDefault)
-			case *roachpb.StoreNotFoundError:
-				log.Warningf(ctx, "raft error: node %d claims to not contain store %d for replica %s: %s",
-					resp.FromReplica.NodeID, resp.FromReplica.StoreID, resp.FromReplica, val)
-				return val.GetDetail() // close Raft connection
-			default:
-				log.Warningf(ctx, "got error from r%d, replica %s: %s",
-					resp.RangeID, resp.FromReplica, val)
-			}
-		default:
-			log.Warningf(ctx, "got unknown raft response type %T from replica %s: %s", val, resp.FromReplica, val)
-		}
-		return nil
-	})
-}
-
-// enqueueRaftUpdateCheck asynchronously registers the given range ID to be
-// checked for raft updates when the processRaft goroutine is idle.
-func (s *Store) enqueueRaftUpdateCheck(rangeID roachpb.RangeID) {
-	s.scheduler.EnqueueRaftReady(rangeID)
-}
-
-func (s *Store) processRequestQueue(ctx context.Context, rangeID roachpb.RangeID) {
-	value, ok := s.replicaQueues.Load(int64(rangeID))
-	if !ok {
-		return
-	}
-	q := (*raftRequestQueue)(value)
-	q.Lock()
-	infos := q.infos
-	q.infos = nil
-	q.Unlock()
-
-	var lastRepl *Replica
-	var hadError bool
-	for i, info := range infos {
-		last := i == len(infos)-1
-		pErr := s.withReplicaForRequest(ctx, info.req, func(ctx context.Context, r *Replica) *roachpb.Error {
-			// Save the last Replica we see, since we don't know in advance which
-			// requests will fail during Replica retrieval. We want this later
-			// so we can handle the Raft Ready state all at once.
-			lastRepl = r
-			pErr := s.processRaftRequestWithReplica(ctx, r, info.req)
-			if last {
-				// If this is the last request, we can handle raft.Ready without
-				// giving up the lock. Set lastRepl to nil, so we don't handle it
-				// down below as well.
-				lastRepl = nil
-				_, expl, err := r.handleRaftReadyRaftMuLocked(ctx, noSnap)
-				maybeFatalOnRaftReadyErr(ctx, expl, err)
-			}
-			return pErr
-		})
-		if pErr != nil {
-			hadError = true
-			if err := info.respStream.Send(newRaftMessageResponse(info.req, pErr)); err != nil {
-				// Seems excessive to log this on every occurrence as the other side
-				// might have closed.
-				log.VEventf(ctx, 1, "error sending error: %s", err)
-			}
-		}
-	}
-
-	if hadError {
-		// If we're unable to process a request, consider dropping the request queue
-		// to free up space in the map.
-		// This is relevant if requests failed because the target replica could not
-		// be created (for example due to the Raft tombstone). The particular code
-		// here takes into account that we don't want to drop the queue if there
-		// are other messages waiting on it, or if the target replica exists. Raft
-		// tolerates the occasional dropped message, but our unit tests are less
-		// forgiving.
-		//
-		// See https://github.com/cockroachdb/cockroach/issues/30951#issuecomment-428010411.
-		if _, exists := s.mu.replicas.Load(int64(rangeID)); !exists {
-			q.Lock()
-			if len(q.infos) == 0 {
-				s.replicaQueues.Delete(int64(rangeID))
-			}
-			q.Unlock()
-		}
-	}
-
-	// If lastRepl is not nil, that means that some of the requests succeeded during
-	// Replica retrieval (withReplicaForRequest) but that the last request did not,
-	// otherwise we would have handled this above and set lastRepl to nil.
-	if lastRepl != nil {
-		// lastRepl will be unlocked when we exit withReplicaForRequest above.
-		// It's fine to relock it here (by calling handleRaftReady instead of
-		// handleRaftReadyRaftMuLocked) since racing to handle Raft Ready won't
-		// have any undesirable results.
-		ctx = lastRepl.AnnotateCtx(ctx)
-		_, expl, err := lastRepl.handleRaftReady(ctx, noSnap)
-		maybeFatalOnRaftReadyErr(ctx, expl, err)
-	}
-}
-
-func (s *Store) processReady(ctx context.Context, rangeID roachpb.RangeID) {
-	value, ok := s.mu.replicas.Load(int64(rangeID))
-	if !ok {
-		return
-	}
-
-	r := (*Replica)(value)
-	ctx = r.AnnotateCtx(ctx)
-	start := timeutil.Now()
-	stats, expl, err := r.handleRaftReady(ctx, noSnap)
-	removed := maybeFatalOnRaftReadyErr(ctx, expl, err)
-	elapsed := timeutil.Since(start)
-	s.metrics.RaftWorkingDurationNanos.Inc(elapsed.Nanoseconds())
-	// Warn if Raft processing took too long. We use the same duration as we
-	// use for warning about excessive raft mutex lock hold times. Long
-	// processing time means we'll have starved local replicas of ticks and
-	// remote replicas will likely start campaigning.
-	if elapsed >= defaultReplicaRaftMuWarnThreshold {
-		log.Warningf(ctx, "handle raft ready: %.1fs [applied=%d, batches=%d, state_assertions=%d]",
-			elapsed.Seconds(), stats.entriesProcessed, stats.batchesProcessed, stats.stateAssertions)
-	}
-	if !removed && !r.IsInitialized() {
-		// Only an uninitialized replica can have a placeholder since, by
-		// definition, an initialized replica will be present in the
-		// replicasByKey map. While the replica will usually consume the
-		// placeholder itself, that isn't guaranteed and so this invocation
-		// here is crucial (i.e. don't remove it).
-		//
-		// We need to hold raftMu here to prevent removing a placeholder that is
-		// actively being used by Store.processRaftRequest.
-		r.raftMu.Lock()
-		if s.removePlaceholder(ctx, r.RangeID) {
-			atomic.AddInt32(&s.counts.droppedPlaceholders, 1)
-		}
-		r.raftMu.Unlock()
-	}
-}
-
-func (s *Store) processTick(ctx context.Context, rangeID roachpb.RangeID) bool {
-	value, ok := s.mu.replicas.Load(int64(rangeID))
-	if !ok {
-		return false
-	}
-	livenessMap, _ := s.livenessMap.Load().(IsLiveMap)
-
-	start := timeutil.Now()
-	r := (*Replica)(value)
-	exists, err := r.tick(livenessMap)
-	if err != nil {
-		log.Error(ctx, err)
-	}
-	s.metrics.RaftTickingDurationNanos.Inc(timeutil.Since(start).Nanoseconds())
-	return exists // ready
-}
-
-// nodeIsLiveCallback is invoked when a node transitions from non-live
-// to live. Iterate through all replicas and find any which belong to
-// ranges containing the implicated node. Unquiesce if currently
-// quiesced. Note that this mechanism can race with concurrent
-// invocations of processTick, which may have a copy of the previous
-// livenessMap where the now-live node is down. Those instances should
-// be rare, however, and we expect the newly live node to eventually
-// unquiesce the range.
-func (s *Store) nodeIsLiveCallback(nodeID roachpb.NodeID) {
-	// Update the liveness map.
-	s.livenessMap.Store(s.cfg.NodeLiveness.GetIsLiveMap())
-
-	s.mu.replicas.Range(func(k int64, v unsafe.Pointer) bool {
-		r := (*Replica)(v)
-		for _, rep := range r.Desc().Replicas().All() {
-			if rep.NodeID == nodeID {
-				r.unquiesce()
-			}
-		}
-		return true
-	})
-}
-
-func (s *Store) processRaft(ctx context.Context) {
-	if s.cfg.TestingKnobs.DisableProcessRaft {
-		return
-	}
-
-	s.scheduler.Start(ctx, s.stopper)
-	// Wait for the scheduler worker goroutines to finish.
-	s.stopper.RunWorker(ctx, s.scheduler.Wait)
-
-	s.stopper.RunWorker(ctx, s.raftTickLoop)
-	s.stopper.RunWorker(ctx, s.coalescedHeartbeatsLoop)
-	s.stopper.AddCloser(stop.CloserFn(func() {
-		s.cfg.Transport.Stop(s.StoreID())
-	}))
-}
-
-func (s *Store) raftTickLoop(ctx context.Context) {
-	ticker := time.NewTicker(s.cfg.RaftTickInterval)
-	defer ticker.Stop()
-
-	var rangeIDs []roachpb.RangeID
-
-	for {
-		select {
-		case <-ticker.C:
-			rangeIDs = rangeIDs[:0]
-			// Update the liveness map.
-			if s.cfg.NodeLiveness != nil {
-				nextMap := s.cfg.NodeLiveness.GetIsLiveMap()
-				for nodeID, entry := range nextMap {
-					if entry.IsLive {
-						// Make sure we ask all live nodes for closed timestamp updates.
-						s.cfg.ClosedTimestamp.Clients.EnsureClient(nodeID)
-						continue
-					}
-					// Liveness claims that this node is down, but ConnHealth gets the last say
-					// because we'd rather quiesce a range too little than one too often.
-					//
-					// NB: This has false negatives. If a node doesn't have a conn open to it
-					// when ConnHealth is called, then ConnHealth will return
-					// rpc.ErrNotHeartbeated regardless of whether the node is up or not. That
-					// said, for the nodes that matter, we're likely talking to them via the
-					// Raft transport, so ConnHealth should usually indicate a real problem if
-					// it gives us an error back. The check can also have false positives if the
-					// node goes down after populating the map, but that matters even less.
-					entry.IsLive = (s.cfg.NodeDialer.ConnHealth(nodeID, rpc.SystemClass) == nil)
-					nextMap[nodeID] = entry
-				}
-				s.livenessMap.Store(nextMap)
-			}
-
-			s.unquiescedReplicas.Lock()
-			// Why do we bother to ever queue a Replica on the Raft scheduler for
-			// tick processing? Couldn't we just call Replica.tick() here? Yes, but
-			// then a single bad/slow Replica can disrupt tick processing for every
-			// Replica on the store which cascades into Raft elections and more
-			// disruption.
-			for rangeID := range s.unquiescedReplicas.m {
-				rangeIDs = append(rangeIDs, rangeID)
-			}
-			s.unquiescedReplicas.Unlock()
-
-			s.scheduler.EnqueueRaftTick(rangeIDs...)
-			s.metrics.RaftTicks.Inc(1)
-
-		case <-s.stopper.ShouldStop():
-			return
-		}
-	}
-}
-
-// Since coalesced heartbeats adds latency to heartbeat messages, it is
-// beneficial to have it run on a faster cycle than once per tick, so that
-// the delay does not impact latency-sensitive features such as quiescence.
-func (s *Store) coalescedHeartbeatsLoop(ctx context.Context) {
-	ticker := time.NewTicker(s.cfg.CoalescedHeartbeatsInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ticker.C:
-			s.sendQueuedHeartbeats(ctx)
-		case <-s.stopper.ShouldStop():
-			return
-		}
-	}
-}
-
-// sendQueuedHeartbeatsToNode requires that the s.coalescedMu lock is held. It
-// returns the number of heartbeats that were sent.
-func (s *Store) sendQueuedHeartbeatsToNode(
-	ctx context.Context, beats, resps []RaftHeartbeat, to roachpb.StoreIdent,
-) int {
-	var msgType raftpb.MessageType
-
-	if len(beats) == 0 && len(resps) == 0 {
-		return 0
-	} else if len(resps) == 0 {
-		msgType = raftpb.MsgHeartbeat
-	} else if len(beats) == 0 {
-		msgType = raftpb.MsgHeartbeatResp
-	} else {
-		log.Fatal(ctx, "cannot coalesce both heartbeats and responses")
-	}
-
-	chReq := &RaftMessageRequest{
-		RangeID: 0,
-		ToReplica: roachpb.ReplicaDescriptor{
-			NodeID:    to.NodeID,
-			StoreID:   to.StoreID,
-			ReplicaID: 0,
-		},
-		FromReplica: roachpb.ReplicaDescriptor{
-			NodeID:  s.Ident.NodeID,
-			StoreID: s.Ident.StoreID,
-		},
-		Message: raftpb.Message{
-			Type: msgType,
-		},
-		Heartbeats:     beats,
-		HeartbeatResps: resps,
-	}
-
-	if log.V(4) {
-		log.Infof(ctx, "sending raft request (coalesced) %+v", chReq)
-	}
-
-	if !s.cfg.Transport.SendAsync(chReq, rpc.SystemClass) {
-		for _, beat := range beats {
-			if value, ok := s.mu.replicas.Load(int64(beat.RangeID)); ok {
-				(*Replica)(value).addUnreachableRemoteReplica(beat.ToReplicaID)
-			}
-		}
-		for _, resp := range resps {
-			if value, ok := s.mu.replicas.Load(int64(resp.RangeID)); ok {
-				(*Replica)(value).addUnreachableRemoteReplica(resp.ToReplicaID)
-			}
-		}
-		return 0
-	}
-	return len(beats) + len(resps)
-}
-
-func (s *Store) sendQueuedHeartbeats(ctx context.Context) {
-	s.coalescedMu.Lock()
-	heartbeats := s.coalescedMu.heartbeats
-	heartbeatResponses := s.coalescedMu.heartbeatResponses
-	s.coalescedMu.heartbeats = map[roachpb.StoreIdent][]RaftHeartbeat{}
-	s.coalescedMu.heartbeatResponses = map[roachpb.StoreIdent][]RaftHeartbeat{}
-	s.coalescedMu.Unlock()
-
-	var beatsSent int
-
-	for to, beats := range heartbeats {
-		beatsSent += s.sendQueuedHeartbeatsToNode(ctx, beats, nil, to)
-	}
-	for to, resps := range heartbeatResponses {
-		beatsSent += s.sendQueuedHeartbeatsToNode(ctx, nil, resps, to)
-	}
-	s.metrics.RaftCoalescedHeartbeatsPending.Update(int64(beatsSent))
-}
-
-var errRetry = errors.New("retry: orphaned replica")
-
-// getOrCreateReplica returns a replica for the given RangeID, creating an
-// uninitialized replica if necessary. The caller must not hold the store's
-// lock. The returned replica has Replica.raftMu locked and it is the caller's
-// responsibility to unlock it.
-func (s *Store) getOrCreateReplica(
-	ctx context.Context,
-	rangeID roachpb.RangeID,
-	replicaID roachpb.ReplicaID,
-	creatingReplica *roachpb.ReplicaDescriptor,
-	isLearner bool,
-) (_ *Replica, created bool, _ error) {
-	// We need a retry loop as the replica we find in the map may be in the
-	// process of being removed or may need to be removed. Retries in the loop
-	// imply that a removal is actually being carried out, not that we're waiting
-	// on a queue.
-	r := retry.Start(retry.Options{
-		InitialBackoff: time.Microsecond,
-		// Set the backoff up to only a small amount to wait for data that
-		// might need to be cleared.
-		MaxBackoff: 10 * time.Millisecond,
-	})
-	for {
-		r.Next()
-		r, created, err := s.tryGetOrCreateReplica(
-			ctx,
-			rangeID,
-			replicaID,
-			creatingReplica,
-			isLearner,
-		)
-		if err == errRetry {
-			continue
-		}
-		if err != nil {
-			return nil, false, err
-		}
-		return r, created, err
-	}
-}
-
-// tryGetOrCreateReplica performs a single attempt at trying to lookup or
-// create a replica. It will fail with errRetry if it finds a Replica that has
-// been destroyed (and is no longer in Store.mu.replicas) or if during creation
-// another goroutine gets there first. In either case, a subsequent call to
-// tryGetOrCreateReplica will likely succeed, hence the loop in
-// getOrCreateReplica.
-func (s *Store) tryGetOrCreateReplica(
-	ctx context.Context,
-	rangeID roachpb.RangeID,
-	replicaID roachpb.ReplicaID,
-	creatingReplica *roachpb.ReplicaDescriptor,
-	isLearner bool,
-) (_ *Replica, created bool, _ error) {
-	// The common case: look up an existing (initialized) replica.
-	if value, ok := s.mu.replicas.Load(int64(rangeID)); ok {
-		repl := (*Replica)(value)
-		repl.raftMu.Lock() // not unlocked on success
-		repl.mu.Lock()
-
-		// Drop messages from replicas we know to be too old.
-		if fromReplicaIsTooOld(repl, creatingReplica) {
-			repl.mu.Unlock()
-			repl.raftMu.Unlock()
-			return nil, false, roachpb.NewReplicaTooOldError(creatingReplica.ReplicaID)
-		}
-
-		// The current replica is removed, go back around.
-		if repl.mu.destroyStatus.Removed() {
-			repl.mu.Unlock()
-			repl.raftMu.Unlock()
-			return nil, false, errRetry
-		}
-
-		toTooOld := toReplicaIsTooOld(repl, replicaID)
-		isPreemptiveSnapshot := repl.mu.replicaID == 0 && repl.isInitializedRLocked()
-		// We need to remove preemptive snapshots when we determine that we're now a
-		// learner. Otherwise we risk appending and applying log entries while we're
-		// not a member of the range and potentially applying a merge which would be
-		// unsafe. See the comment in Replica.acquireMergeLock().
-		removePreemptiveSnapshot := isPreemptiveSnapshot && isLearner
-		// The current replica needs to be removed, remove it and go back around.
-		if toTooOld || removePreemptiveSnapshot {
-
-			if shouldLog := log.V(1); shouldLog && toTooOld {
-				log.Infof(ctx, "found message for replica ID %d which is newer than %v",
-					replicaID, repl)
-			} else if shouldLog && removePreemptiveSnapshot {
-				log.Infof(ctx, "found message for replica ID %v as non-voter but "+
-					"currently not part of the range, destroying preemptive snapshot",
-					replicaID)
-			}
-
-			repl.mu.Unlock()
-			if err := s.removeReplicaRaftMuLocked(ctx, repl, replicaID, RemoveOptions{
-				DestroyData: true,
-			}); err != nil {
-				log.Fatalf(ctx, "failed to remove replica: %v", err)
-			}
-			repl.raftMu.Unlock()
-			return nil, false, errRetry
-		}
-		defer repl.mu.Unlock()
-
-		// If this is intended for replicaID 0 then it's either a preemptive
-		// snapshot or a split/merge lock in which case we'll let it go through.
-		if replicaID == 0 {
-			return repl, false, nil
-		}
-		var err error
-		if repl.mu.replicaID == 0 {
-			// This message is telling us about our replica ID.
-			// This is a common case when dealing with preemptive snapshots.
-			err = repl.setReplicaIDRaftMuLockedMuLocked(repl.AnnotateCtx(ctx), replicaID)
-		} else if repl.mu.replicaID > replicaID {
-			// The sender is behind and is sending to an old replica.
-			// We could silently drop this message but this way we'll inform the
-			// sender that they may no longer exist.
-			err = roachpb.NewRangeNotFoundError(rangeID, s.StoreID())
-		} else if repl.mu.replicaID != replicaID {
-			// This case should have been caught by handleToReplicaTooOld.
-			log.Fatalf(ctx, "intended replica id %d unexpectedly does not match the current replica %v",
-				replicaID, repl)
-		}
-		if err != nil {
-			repl.raftMu.Unlock()
-			return nil, false, err
-		}
-		return repl, false, nil
-	}
-
-	// No replica currently exists, so we'll try to create one. Before creating
-	// the replica, see if there is a tombstone which would indicate that this
-	// is a stale message.
-	// NB: we check this before creating a new Replica and adding it to the
-	// Store's Range map even though we must check it again after to avoid race
-	// conditions. This double-checked locking is an optimization to avoid this
-	// work when we know the Replica should not be created ahead of time.
-	tombstoneKey := keys.RaftTombstoneKey(rangeID)
-	var tombstone roachpb.RaftTombstone
-	if ok, err := engine.MVCCGetProto(
-		ctx, s.Engine(), tombstoneKey, hlc.Timestamp{}, &tombstone, engine.MVCCGetOptions{},
-	); err != nil {
-		return nil, false, err
-	} else if ok && replicaID != 0 && replicaID < tombstone.NextReplicaID {
-		return nil, false, &roachpb.RaftGroupDeletedError{}
-	}
-
-	// Create a new replica and lock it for raft processing.
-	repl := newReplica(rangeID, s)
-	repl.creatingReplica = creatingReplica
-	repl.raftMu.Lock() // not unlocked
-
-	// Install the replica in the store's replica map. The replica is in an
-	// inconsistent state, but nobody will be accessing it while we hold its
-	// locks.
-	s.mu.Lock()
-	// Grab the internal Replica state lock to ensure nobody mucks with our
-	// replica even outside of raft processing. Have to do this after grabbing
-	// Store.mu to maintain lock ordering invariant.
-	repl.mu.Lock()
-	repl.mu.tombstoneMinReplicaID = tombstone.NextReplicaID
-	// Add the range to range map, but not replicasByKey since the range's start
-	// key is unknown. The range will be added to replicasByKey later when a
-	// snapshot is applied. After unlocking Store.mu above, another goroutine
-	// might have snuck in and created the replica, so we retry on error.
-	if err := s.addReplicaToRangeMapLocked(repl); err != nil {
-		repl.mu.Unlock()
-		s.mu.Unlock()
-		repl.raftMu.Unlock()
-		return nil, false, errRetry
-	}
-	s.mu.uninitReplicas[repl.RangeID] = repl
-	s.mu.Unlock() // NB: unlocking out of order
-
-	// Initialize the Replica with the replicaID.
-	if err := func() error {
-		// Check for a tombstone again now that we've inserted into the Range
-		// map. This double-checked locking ensures that we avoid a race where a
-		// replica is created and destroyed between the initial unsynchronized
-		// tombstone check and the Range map linearization point. By checking
-		// again now, we make sure to synchronize with any goroutine that wrote
-		// a tombstone and then removed an old replica from the Range map.
-		if ok, err := engine.MVCCGetProto(
-			ctx, s.Engine(), tombstoneKey, hlc.Timestamp{}, &tombstone, engine.MVCCGetOptions{},
-		); err != nil {
-			return err
-		} else if ok && replicaID != 0 && replicaID < tombstone.NextReplicaID {
-			return &roachpb.RaftGroupDeletedError{}
-		}
-
-		// An uninitialized replica should have an empty HardState.Commit at
-		// all times. Failure to maintain this invariant indicates corruption.
-		// And yet, we have observed this in the wild. See #40213.
-		if hs, err := repl.mu.stateLoader.LoadHardState(ctx, s.Engine()); err != nil {
-			return err
-		} else if hs.Commit != 0 {
-			log.Fatalf(ctx, "found non-zero HardState.Commit on uninitialized replica %s. HS=%+v", repl, hs)
-		}
-
-		desc := &roachpb.RangeDescriptor{
-			RangeID: rangeID,
-			// NB: other fields are unknown; need to populate them from
-			// snapshot.
-		}
-		return repl.initRaftMuLockedReplicaMuLocked(desc, s.Clock(), replicaID)
-	}(); err != nil {
-		// Mark the replica as destroyed and remove it from the replicas maps to
-		// ensure nobody tries to use it
-		repl.mu.destroyStatus.Set(errors.Wrapf(err, "%s: failed to initialize", repl), destroyReasonRemoved)
-		repl.mu.Unlock()
-		s.mu.Lock()
-		s.unlinkReplicaByRangeIDLocked(rangeID)
-		s.mu.Unlock()
-		repl.raftMu.Unlock()
-		return nil, false, err
-	}
-	repl.mu.Unlock()
-	return repl, true, nil
-}
-
-// isFromReplicaTooOld returns an true if the creatingReplica is deemed to be
-// a member of the range which has been removed.
-// Assumes toReplica.mu is held.
-func fromReplicaIsTooOld(toReplica *Replica, fromReplica *roachpb.ReplicaDescriptor) bool {
-	toReplica.mu.AssertHeld()
-	if fromReplica == nil {
-		return false
-	}
-	desc := toReplica.mu.state.Desc
-	_, found := desc.GetReplicaDescriptorByID(fromReplica.ReplicaID)
-	return !found && fromReplica.ReplicaID < desc.NextReplicaID
-}
-
-// toReplicaIsTooOld returns true if replicaID is newer than toReplica
-// indicating that the Replica needs to be removed.
-// Assumes toReplica.mu is held.
-func toReplicaIsTooOld(toReplica *Replica, replicaID roachpb.ReplicaID) bool {
-	toReplica.mu.AssertHeld()
-	return replicaID != 0 && toReplica.mu.replicaID != 0 &&
-		toReplica.mu.replicaID < replicaID
-}
-
-func (s *Store) updateCapacityGauges() error {
-	desc, err := s.Descriptor(false /* useCached */)
-	if err != nil {
-		return err
-	}
-	s.metrics.Capacity.Update(desc.Capacity.Capacity)
-	s.metrics.Available.Update(desc.Capacity.Available)
-	s.metrics.Used.Update(desc.Capacity.Used)
-
-	return nil
 }
 
 // updateReplicationGauges counts a number of simple replication statistics for
